finalize integration tests

+++ src/main/java/com/outbrain/test/api/PostService.java
@@ -28,13 +28,13 @@ public class PostService {
         return reply;
     }
 
-    public Post getPost(int id) {
+    public Post getPost(long id) {
         LOG.debug("Fetching post with id: {}", id);
         com.outbrain.test.dal.Post post = repo.getPost(id);
         return new Post(post);
     }
 
-    public Post deletePost(int id) {
+    public Post deletePost(long id) {
         LOG.debug("Deleting post with id: {}", id);
         com.outbrain.test.dal.Post post = repo.deletePost(id);
         if (post == null) {
@@ -49,7 +49,7 @@ public class PostService {
             com.outbrain.test.dal.Post newPost = repo.addPost(post.title, post.content, post.author);
             return new Post(newPost);
         } else {
-            com.outbrain.test.dal.Post existing = repo.modifyPost(Integer.parseInt(post.id), post.title, post.content, post.author);
+            com.outbrain.test.dal.Post existing = repo.modifyPost(Long.parseLong(post.id), post.title, post.content, post.author);
             return new Post(existing);
         }
     }
